response status exception first commit

+++ src/main/java/com/learnwithranjan/restservices/services/UserServices.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.learnwithranjan.restservices.entities.User;
+import com.learnwithranjan.restservices.exceptions.UserExistsException;
+import com.learnwithranjan.restservices.exceptions.UserNotFound;
 import com.learnwithranjan.restservices.repositories.UserRepository;
 
 @Service
@@ -20,31 +22,70 @@ public class UserServices {
 		return userRepository.findAll();
 	}
 	
-	public User createUsers(User user) {
+	//without handle exception
+	/*public User createUsers(User user) {
+		return userRepository.save(user);
+	}*/
+	//handle exception
+	public User createUsers(User user) throws UserExistsException {
+		User existUser = userRepository.findByUsername(user.getUsername());
+		if(existUser != null) {
+			throw new UserExistsException("User alredy exists in repository");
+		}
 		return userRepository.save(user);
 	}
 	
 	
-	public Optional<User> getUserByID(Long id) {
+	// without handle exception 
+	/*public Optional<User> getUserByID(Long id) {
 		Optional<User> user = userRepository.findById(id);
+		return user;
+	}*/
+	
+	//handle exception
+	public Optional<User> getUserByID(Long id) throws UserNotFound {
+		Optional<User> user = userRepository.findById(id);
+		if(!user.isPresent()) {
+			throw new UserNotFound("User Not found in repository");
+		}
 		return user;
 	}
 	
-	
-	public User updateUserByID(Long id,User user) {
+	//without handle exception
+	/*public User updateUserByID(Long id,User user) {
 		user.setId(id);
+		return userRepository.save(user);
+	}*/
+	
+	//handle exception
+	public User updateUserByID(Long id,User user) throws UserNotFound {
+		Optional<User> optionalUser = userRepository.findById(id);
+		if(!optionalUser.isPresent()) {
+			throw new UserNotFound("User Not found in repository, provide correct repository");
+		}
 		return userRepository.save(user);
 	}
 	
+	//without handle exception
+	/*public void deleteUserByID(Long id) {
+		if(userRepository.findById(id).isPresent()) {
+			userRepository.deleteById(id);
+		}
+	}*/
 	
-	public void deleteUserByID(Long id) {
+	//handle exception
+	public void deleteUserByID(Long id) throws UserNotFound {
+		Optional<User> optionalUser = userRepository.findById(id);
+		if(!optionalUser.isPresent()) {
+			throw new UserNotFound("User Not found in repository, provide correct user repository");
+		}
 		if(userRepository.findById(id).isPresent()) {
 			userRepository.deleteById(id);
 		}
 	}
 	
 	public User getUserByUserName(String userName) {
-		return userRepository.getUserByUsername(userName);
+		return userRepository.findByUsername(userName);
 	}
 
 }
